refactor(src): Refactored the packagestructure so it correlates with the given case
Refactored the packagestructure to represent entity, logic and model. I
also added utils package for "helper" classes that will be added later
such as enums, validation classes, filehandling, logger etc.

Closed issue #27
