+++ src/main/java/com/ntnu/idatt2003/entity/Player.java
@@ -1,4 +1,4 @@
-package com.ntnu.idatt2003.core;
+package com.ntnu.idatt2003.entity;
 
 /*
 import org.slf4j.Logger;
@@ -6,7 +6,8 @@ import org.slf4j.LoggerFactory;
  */
 
 
-import com.ntnu.idatt2003.tile.Tile;
+import com.ntnu.idatt2003.logic.BoardGame;
+import com.ntnu.idatt2003.model.Tile;
 
 /**
  * The type Player.
