+++ src/main/java/com/ntnu/idatt2003/logic/action/TileAction.java
@@ -1,0 +1,8 @@
+package com.ntnu.idatt2003.logic.action;
+
+import com.ntnu.idatt2003.entity.Player;
+
+public interface TileAction {
+  
+  void perform(Player player);
+}
