+++ src/main/java/com/ntnu/idatt2003/model/Board.java
@@ -1,6 +1,5 @@
-package com.ntnu.idatt2003.core;
+package com.ntnu.idatt2003.model;
 
-import com.ntnu.idatt2003.tile.Tile;
 import java.util.HashMap;
 import java.util.Map;
