+++ src/main/java/com/ntnu/idatt2003/model/Dice.java
@@ -1,5 +1,6 @@
-package com.ntnu.idatt2003.core;
+package com.ntnu.idatt2003.model;
 
+import com.ntnu.idatt2003.entity.Die;
 import java.util.ArrayList;
 import java.util.List;
