--- src/main/java/com/ntnu/idatt2003/tile/action/TileAction.java
@@ -1,8 +1,0 @@
-package com.ntnu.idatt2003.tile.action;
-
-import com.ntnu.idatt2003.core.Player;
-
-public interface TileAction {
-  
-  void perform(Player player);
-}
