+++ src/main/java/com/ntnu/idatt2003/view/MainApplication.java
@@ -1,4 +1,4 @@
-package com.ntnu.idatt2003;
+package com.ntnu.idatt2003.view;
 
 import javafx.application.Application;
 import javafx.stage.Stage;
